changed the timing array from int32 to int16 to reduce memory, also trying to use no allocate cache user mase probably 0x30F

+++ timings/parse.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 import numpy,sys
 
-a=numpy.fromfile(sys.argv[1],dtype=numpy.int32,count=2*int(sys.argv[2]))
+a=numpy.fromfile(sys.argv[1],dtype=numpy.int16,count=2*int(sys.argv[2]))
 for i in range(2,int(sys.argv[2])):
 	print (a[2*i+1]-a[2*i])
